Change access modifier of SlackApp member variables
This commit updates the access level of a number of members in the SlackApp class to be protected. This allows subclasses of SlackApp to directly access and manipulate these members, increasing the flexibility of the class for subclassing. The affected members include slashCommands, events, globalShorcuts, messageShorcuts, blockActions, blockSuggestions, viewSubmissions, and viewClosed.